requirements.txt
python-dotenv
paramiko
jinja2